requirements.txt
torch
transformers>=4.40.0
sentence-transformers>=2.6.0
faiss-cpu
scikit-learn
numpy
pyyaml
nltk
